requirements.txt
Django==4.1.3
django_ckeditor==6.5.1
django_extensions==3.2.1
